[UPDATE] move swagger for directory config
